mini-course-builder/requirements.txt
Django==4.2.23
djangorestframework==3.16.0
django-cors-headers==4.7.0
openai==1.86.0
python-dotenv==1.1.0
weasyprint==65.1 